backend/requirements.txt
albucore==0.0.23
albumentations==2.0.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Cython==3.0.12
decorator==5.2.1
Django==5.1.7
django-allauth==65.6.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
fire==0.7.0
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
jose==1.0.0
kombu==5.5.2
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
networkx==3.4.2
nulltype==2.3.1
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pillow==11.1.0
plaid-python==29.1.0
prompt_toolkit==3.0.50
protobuf==6.30.2
pyasn1==0.4.8
pyclipper==1.3.0.post6
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
redis==5.2.1
requests==2.32.3
rsa==4.9
scikit-image==0.25.2
scipy==1.15.2
setuptools==78.1.0
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stringzilla==3.12.3
termcolor==3.0.1
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
